added editUser and RemoveUser in UserController and moved registerUser to UserController, fixed routes
